replace RNN lava apply
